backend/requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cors==1.0.1
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
PySocks==1.7.1
python-dotenv==0.18.0
requests==2.26.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==1.4.21
tldextract==3.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
validate-docbr==1.8.2
validate-email==1.3
Werkzeug==2.0.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
